# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "operationId": "GroupsController_list",
        "summary": "List groups",
        "description": "Filterable by branchId and status. Requires group:read.",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_create",
        "summary": "Create a group",
        "description": "Requires group:manage.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}": {
      "get": {
        "operationId": "GroupsController_get",
        "summary": "Get a group by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_update",
        "summary": "Update a group",
        "description": "Updates a group's own fields (name, capacity, etc). Requires group:manage.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/children": {
      "get": {
        "operationId": "GroupsController_children",
        "summary": "List a group's children",
        "description": "Returns children currently assigned to this group. Requires group:read.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupChildDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_assignChild",
        "summary": "Assign a child to a group",
        "description": "Closes out the child's current group assignment (if any) and opens a new one in this group from dto.effectiveDate. Rejected with CAPACITY_EXCEEDED if the group is full unless ?force=true — and force=true itself requires the group:capacity_override permission on top of group:manage.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignChildDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupChildDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/transfer": {
      "post": {
        "operationId": "GroupsController_transfer",
        "summary": "Transfer a child to another group",
        "description": "Same capacity check and force/group:capacity_override rule as assign — the destination group is dto.toGroupId, not the :id in the URL (which is informational only).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferChildDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/staff": {
      "get": {
        "operationId": "GroupsController_staff",
        "summary": "List a group's staff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStaffDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "put": {
        "operationId": "GroupsController_setStaff",
        "summary": "Set a group's staff",
        "description": "Full replace (PUT) of the group's assigned staff list. Requires group:manage.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/history": {
      "get": {
        "operationId": "GroupsController_history",
        "summary": "Get a group's assignment history",
        "description": "Every child assignment ever made to this group, most recent first.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kindergarten API",
    "description": "Kindergarten Management System — backend API (Stage 1)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Quyoshcha"
          },
          "ageMinMonths": {
            "type": "object",
            "nullable": true,
            "example": 36
          },
          "ageMaxMonths": {
            "type": "object",
            "nullable": true,
            "example": 48
          },
          "capacity": {
            "type": "number",
            "example": 20
          },
          "workingHours": {
            "type": "object",
            "nullable": true,
            "description": "e.g. { \"open\": \"07:30\", \"close\": \"19:00\" }"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "currentCount": {
            "type": "number",
            "description": "Children currently assigned (effectiveTo = null)"
          },
          "availablePlaces": {
            "type": "number",
            "description": "capacity − currentCount"
          }
        },
        "required": [
          "id",
          "branchId",
          "name",
          "capacity",
          "status",
          "currentCount",
          "availablePlaces"
        ]
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ageMinMonths": {
            "type": "number"
          },
          "ageMaxMonths": {
            "type": "number"
          },
          "capacity": {
            "type": "number",
            "default": 20
          },
          "workingHours": {
            "type": "object",
            "example": {
              "open": "07:00",
              "close": "19:00"
            }
          }
        },
        "required": [
          "branchId",
          "name"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ageMinMonths": {
            "type": "number"
          },
          "ageMaxMonths": {
            "type": "number"
          },
          "capacity": {
            "type": "number",
            "default": 20
          },
          "workingHours": {
            "type": "object",
            "example": {
              "open": "07:00",
              "close": "19:00"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          }
        }
      },
      "GroupChildDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "birthDate",
          "status"
        ]
      },
      "AssignChildDto": {
        "type": "object",
        "properties": {
          "childId": {
            "type": "string"
          },
          "effectiveDate": {
            "type": "string"
          }
        },
        "required": [
          "childId",
          "effectiveDate"
        ]
      },
      "TransferChildDto": {
        "type": "object",
        "properties": {
          "childId": {
            "type": "string"
          },
          "toGroupId": {
            "type": "string"
          },
          "effectiveDate": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "childId",
          "toGroupId",
          "effectiveDate"
        ]
      },
      "StaffMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "description": "Staff DO have fullName — children do not"
          }
        },
        "required": [
          "id",
          "fullName"
        ]
      },
      "GroupStaffDto": {
        "type": "object",
        "properties": {
          "mainTeacher": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StaffMemberDto"
              }
            ]
          },
          "assistant": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StaffMemberDto"
              }
            ]
          },
          "nurse": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StaffMemberDto"
              }
            ]
          }
        }
      },
      "SetStaffDto": {
        "type": "object",
        "properties": {
          "mainTeacherId": {
            "type": "string"
          },
          "assistantTeacherId": {
            "type": "string"
          },
          "nurseId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}